+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.58.0-preview] - 2023-09-04
+- PVP: Add x-ray checks `PVP-{33,34}-1` to verify that there are no filenames that ought to never appear in a package. (PVS-186)
+- PVP: Add x-ray check `PVP-{35,82}-1` to verify that there are no unmodified Packager Starter Kit files. (PVS-186)
+- PVP: Add x-ray check `PVP-36-1` to verify that there are no Git LFS pointer files.
+- PVP: Add x-ray check `PVP-37-1` to verify that there are no Unity assets using binary serialization (PETS-1381).
+
 ## [0.57.0-preview] - 2023-08-22
 - Fix invocation of API updater for certain Unity branches. (PVS-185)
 - PVP: Add x-ray check `PVP-155-1` to verify that `Documentation~/filter.yml` is valid. (PVS-184)